refactor: Update hooks and components for client-side rendering and improve session handling
export const dynamic = "force-dynamic"; to avoid pre-rendering for dynamic routes in standalone mode.
