+++ app/auth/signin/page.tsx
@@ -1,7 +1,9 @@
 import { SigninClient } from "./signin-client";
 import { getEnabledProviders } from "@/lib/auth";
 
-export default async function SigninPage() {
+export const dynamic = "force-dynamic";
+
+export default async function Page() {
   const providers = getEnabledProviders();
   return <SigninClient providers={providers} />;
 }
