+++ tests/pages/page.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, screen, fireEvent, act } from "@testing-library/react";
-import Home from "@/app/home/page";
+import Home from "@/app/home/page-client";
 import { useChat } from "@/hooks/use-chat";
 import { useIsMobile } from "@/hooks/use-mobile";
